replace multiproof with imp
