+++ Cargo.toml
@@ -24,15 +24,16 @@ debug = false
 
 [dependencies]
 arrayref = "0.3.5"
-sha2 = "0.8.0"
 bigint = "4.4.1"
 hex = "0.3.2"
+imp = "0.0.1"
 qimalloc = "0.1.0"
+sha2 = "0.8.0"
 
 [dev-dependencies]
+composer = { path = "./composer" }
 criterion = "0.3"
 ewasm = "0.1.2"
-composer = { path = "./composer" }
 
 [[bench]]
 name = "execution"
