+++ composer/Cargo.toml
@@ -10,4 +10,5 @@ bigint = "4.4.1"
 rand = "0.7"
 sha2 = "0.8.0"
 hex = "0.3.2"
+imp = "0.0.1"
 sheth = { path = "../", features = ["std"] }
